refactor: remove old yubikey implemenation
